Combine Fidenza and IC sales channels
Since the channel number is being used as the JSON key, it's not
possible for the same channel to have two entries within the channels
object. To circumvent this, we implement @ryley-o's suggestion to
combine the Fidenza and Incomplete Control sales channels into a single
channel.

We can revisit this later should it be necessary to split the sales
feeds into two different channels.