Converted smartBotResponse.js to typescript
Currently left smartBotResponse.js in the Utils folder so you can
refer back to it if needed. Happy to remove the file and commit again
if needed.

Types were added on variables and functions where it's clear what
the intended type should be. Also exported the smartBotResponse
func and changed the 'dotenv' import to be consistent with other TS
files.